Updated clock frequency adjustment and TIMER1 w/ TIMER2 for counting
The clock PLL register that was used was causing errors in transmitting
information (UART) and causing errors while setting-up both TIMER1 and TIMER2.

A new prescaler was set (N1 = N2 = 2; M = 65) and the UART was adjusted for
that purpose. In addition, the TIMER1 was used for counting microseconds
and TIMER2 for counting milliseconds