Updated 'motor_t' structure with pointer to TMR functions
Now it is possible to know which timer is attached to each
motor so then when it starts moving it also starts the timer
for knowing the exact position it is.

In addition, the timers now handles whether a movement ends,
updating the specific variable so then we know which motor
have ended moving.

This allows the application to continue working and detect
the motor(s) that ended.